♻️ [refactor/withdraw] CKPS-16 탈퇴 사유 폴더 위치 리펙토링
